+++ webook/internal/integration/setup/wire_gen.go
@@ -14,6 +14,7 @@ import (
 	"gitee.com/train-cloud/geektime-basic-go/internal/web"
 	"gitee.com/train-cloud/geektime-basic-go/internal/web/jwt"
 	"gitee.com/train-cloud/geektime-basic-go/ioc"
+	"gitee.com/train-cloud/geektime-basic-go/pkg/streamer"
 	"github.com/gin-gonic/gin"
 	"github.com/google/wire"
 )
@@ -68,7 +69,8 @@ func InitWebServer() *gin.Engine {
 	llmConfig := ioc.InitLLMConfig()
 	llmClient := ioc.InitLLMClient(llmConfig, loggerX)
 	toolExecutor := service.NewAIChatToolExecutor(articleSearchService, articleReaderService, interactionRepository, loggerX)
-	chatService := service.NewAIChatService(conversationRepository, messageRepository, llmClient, articleSearchService, toolExecutor, loggerX)
+	eventStreamer := streamer.NewRedisStreamer(cmdable)
+	chatService := service.NewAIChatService(conversationRepository, messageRepository, llmClient, articleSearchService, toolExecutor, loggerX, eventStreamer)
 	limiter := ioc.InitChatLimiter(cmdable)
 	chatHandler := web.NewInternalChatHandler(chatService, loggerX, limiter)
 	articleSearchHandler := web.NewInternalArticleSearchHandler(articleSearchService, loggerX)
@@ -185,7 +187,8 @@ func InitChatHandler() web.ChatHandler {
 	interactionCache := cache.NewRedisInteractionCache(cmdable)
 	interactionRepository := repository.NewCacheInteractionRepository(interactionDAO, interactionCache, loggerX)
 	toolExecutor := service.NewAIChatToolExecutor(articleSearchService, articleReaderService, interactionRepository, loggerX)
-	chatService := service.NewAIChatService(conversationRepository, messageRepository, llmClient, articleSearchService, toolExecutor, loggerX)
+	eventStreamer := streamer.NewRedisStreamer(cmdable)
+	chatService := service.NewAIChatService(conversationRepository, messageRepository, llmClient, articleSearchService, toolExecutor, loggerX, eventStreamer)
 	limiter := ioc.InitChatLimiter(cmdable)
 	chatHandler := web.NewInternalChatHandler(chatService, loggerX, limiter)
 	return chatHandler
@@ -215,4 +218,4 @@ var clickEventSvcProvider = wire.NewSet(dao.NewGormAIClickEventDAO, cache.NewRed
 
 var polishSvcProvider = wire.NewSet(service.NewAIArticlePolishService)
 
-var chatSvcProvider = wire.NewSet(ioc.InitLLMConfig, ioc.InitLLMClient, ioc.InitChatLimiter, dao.NewGormConversationDAO, dao.NewGormMessageDAO, cache.NewRedisConversationCache, cache.NewRedisMessageCache, repository.NewCacheConversationRepository, repository.NewCacheMessageRepository, service.NewAIChatService, service.NewAIChatToolExecutor)
+var chatSvcProvider = wire.NewSet(ioc.InitLLMConfig, ioc.InitLLMClient, ioc.InitChatLimiter, dao.NewGormConversationDAO, dao.NewGormMessageDAO, cache.NewRedisConversationCache, cache.NewRedisMessageCache, repository.NewCacheConversationRepository, repository.NewCacheMessageRepository, service.NewAIChatService, service.NewAIChatToolExecutor, streamer.NewRedisStreamer)